+++ packages/dd-trace/test/appsec/rasp/index.spec.js
@@ -1,36 +1,10 @@
 'use strict'
 
-const proxyquire = require('proxyquire')
+const rasp = require('../../../src/appsec/rasp')
 const { handleUncaughtExceptionMonitor } = require('../../../src/appsec/rasp')
 
 describe('RASP', () => {
-  let waf, rasp, datadogCore, stackTrace, web
-
   beforeEach(() => {
-    datadogCore = {
-      storage: {
-        getStore: sinon.stub()
-      }
-    }
-    waf = {
-      run: sinon.stub()
-    }
-
-    stackTrace = {
-      reportStackTrace: sinon.stub()
-    }
-
-    web = {
-      root: sinon.stub()
-    }
-
-    rasp = proxyquire('../../../src/appsec/rasp', {
-      '../../../../datadog-core': datadogCore,
-      '../waf': waf,
-      '../stack_trace': stackTrace,
-      './../../plugins/util/web': web
-    })
-
     const config = {
       appsec: {
         stackTrace: {
