+++ packages/dd-trace/test/appsec/rasp/ssrf.express.plugin.spec.js
@@ -6,21 +6,11 @@ const appsec = require('../../../src/appsec')
 const Config = require('../../../src/config')
 const path = require('path')
 const { assert } = require('chai')
+const { checkRaspExecutedAndNotThreat, checkRaspExecutedAndHasThreat } = require('./utils')
 
 function noop () {}
 
 describe('RASP - ssrf', () => {
-  function getWebSpan (traces) {
-    for (const trace of traces) {
-      for (const span of trace) {
-        if (span.type === 'web') {
-          return span
-        }
-      }
-    }
-    throw new Error('web span not found')
-  }
-
   withVersions('express', 'express', expressVersion => {
     let app, server, axios
 
@@ -67,15 +57,8 @@ describe('RASP - ssrf', () => {
           if (!e.response) {
             throw e
           }
-          return await agent.use((traces) => {
-            const span = getWebSpan(traces)
-            assert.property(span.meta, '_dd.appsec.json')
-            assert(span.meta['_dd.appsec.json'].includes('rasp-ssrf-rule-id-1'))
-            assert.equal(span.metrics['_dd.appsec.rasp.rule.eval'], 1)
-            assert(span.metrics['_dd.appsec.rasp.duration'] > 0)
-            assert(span.metrics['_dd.appsec.rasp.duration_ext'] > 0)
-            assert.property(span.meta_struct, '_dd.stack')
-          })
+
+          return checkRaspExecutedAndHasThreat(agent, 'rasp-ssrf-rule-id-1')
         }
 
         assert.fail('Request should be blocked')
@@ -92,11 +75,7 @@ describe('RASP - ssrf', () => {
 
             axios.get('/?host=www.datadoghq.com')
 
-            await agent.use((traces) => {
-              const span = getWebSpan(traces)
-              assert.notProperty(span.meta, '_dd.appsec.json')
-              assert.notProperty(span.meta_struct || {}, '_dd.stack')
-            })
+            return checkRaspExecutedAndNotThreat(agent)
           })
 
           it('Should detect threat doing a GET request', async () => {
@@ -148,10 +127,7 @@ describe('RASP - ssrf', () => {
 
             axios.get('/?host=www.datadoghq.com')
 
-            await agent.use((traces) => {
-              const span = getWebSpan(traces)
-              assert.notProperty(span.meta, '_dd.appsec.json')
-            })
+            return checkRaspExecutedAndNotThreat(agent)
           })
 
           it('Should detect threat doing a GET request', async () => {
@@ -260,15 +236,7 @@ describe('RASP - ssrf', () => {
 
       assert.equal(response.status, 200)
 
-      await agent.use((traces) => {
-        const span = getWebSpan(traces)
-        assert.property(span.meta, '_dd.appsec.json')
-        assert(span.meta['_dd.appsec.json'].includes('rasp-ssrf-rule-id-1'))
-        assert.equal(span.metrics['_dd.appsec.rasp.rule.eval'], 1)
-        assert(span.metrics['_dd.appsec.rasp.duration'] > 0)
-        assert(span.metrics['_dd.appsec.rasp.duration_ext'] > 0)
-        assert.property(span.meta_struct, '_dd.stack')
-      })
+      return checkRaspExecutedAndHasThreat(agent, 'rasp-ssrf-rule-id-1')
     })
   })
 })
